flow: only fake parameter assignments once
This is similar to parsing inlines only once but taking it a bit further.

First of all, all calls should be parsed only once, not just inlines.  And
the inlining code only affected actually parsing the function and inserting
return states into the database.  But we still load the call data into the
DB.

It's kind of tricky to know what things are required and which are not.

But creating fake assignments is not required.  I was worried that the
fake state data was freed too early so I re-worked that a bit.

Signed-off-by: Dan Carpenter <error27@gmail.com>