helper: make strip_expr() strip out "+ 0"
One thing that people have started doing is:

#define foo(x) (COMPILE_TIME_MAGIC + foo_thing(x))

The COMPILE_TIME_MAGIC is something that causes a BUILD_BUG() if conditions
are not met but otherwise ends up being zero.  Zero + x is x.  Except if
the addition causes type promotion and changes the type.  And also in terms
arrays &foo is the same as &foo[0] but they mean different things.

The other thing is that there is some really complicated addition in the
kernel so this change causes a slow down until I added caching.

Signed-off-by: Dan Carpenter <error27@gmail.com>