flow: fix MACRO_ASSIGNMENT_HOOK
This hook isn't used very much.  What it does is that it says if you have
an assignment "foo = MACRO;" then pass the assign expression.  The if
statement is checking that maybe the assignment is inside the MACRO and
we don't want to trigger for that.

So before I was saying if the "=" character is inside the macro then don't
trigger.  But the problem is that for some macro assignments we end up
creating fake assignments.  "foo = frob() ? a : b;"  becomes "foo = a;" and
"foo = b;".  And the location of the "=" is fake and might be inside the
macro depending on how it's faked.  So use the location of "foo" instead
of the location of the "=".

Signed-off-by: Dan Carpenter <error27@gmail.com>