unwind: fix a bug with SSA
I wrote this patch some time back and now I'm not sure at all what I was
thinking.

It looks like we are passing a merged SSA state to a free function and
it's freeing the merged SSA state instead of all the SSA states.  I thought
there was supposed to be a function to change all the states?  Why did I
open code this???

Anyway, it probably silences some false positives so lets merge this crap.

Signed-off-by: Dan Carpenter <error27@gmail.com>