unwind: don't warn about impossible leaks
The issue is that we try to split things between success fail, but there
are also impossible paths which are neither success nor fail.  And if they
get merged with failure paths then the path is now possible but the
resource is not freed on failure.

Handling "&param_released" is special because if it's in &param_released
state then we really want to record that.  Possibly I should move the
parameter tracking code to a different check to make this simpler?

Signed-off-by: Dan Carpenter <error27@gmail.com>