extra: set no limits mark_sub_members_gone()
The issue here is that set_extra_nomod() sets the PARAM_LIMITs.  We only
want to do that for explicit conditions and not for things which let us
know the limit but aren't an explicit check.  For example, dereferencing
a pointer can only succeed if the pointer is non-NULL so we know that the
pointer is non-NULL but it's not an explicit limit check.

Same thing for marking_sub_members_gone().  That's not a limit check at
all.

Signed-off-by: Dan Carpenter <error27@gmail.com>