db/fixup_kernel.sh: ignore param limits from __dev_printk()
The problem is that it turns out PARAM_LIMIT can be pretty expensive to
handle.  It takes time to figure out the implications and it can create
a lot of implied history to parse it.

So with pointer we record a lot that the pointer can be either NULL or
valid on this path.  That seems like useless information and mostly it is.
But sometimes it's really important to know that a pointer can't be an
error pointer.

I am trying to think of a better way to limit this.  There are some very
small limits in is_boring_pointer_info() but I don't have a big picture
solution... :/

What this patch does is that it says ignore dev_err() limit information.
Every error path has a dev_err() on it, so we end up with a ton of
pointless limits saying that the dev->class etc is not an error pointer.
Just manually delete that stuff.

Signed-off-by: Dan Carpenter <error27@gmail.com>