mtag: create a reverse mapping from mtag to symbol
How you create an mtag is you take the name of a function and the scope
and hash it.  That gives you a number.  The number is saved in the DB
etc.

But say you have a number and want to look up the function?  Now you can
use the new get_symbol_from_mtag() function.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>