capped: handle for loops better
The smatch_capped.c stuff only counts variable caps.  So if a value is
capped at 10, that's smatch_extra.c that's not smatch_capped.  This code
says that if either side of the if statement is known then it's not
capped.

But at the start of a for loop the iterator is zero so it's known.  In
other words, this basically ignored for loops, but actually for loops
make things capped.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>