macro_table: don't store duplicate macros
I don't know if this matters...  I had thought that I could use
the get_inner_macro() stuff but it turns out to be useless in practice.
The whole inner/outer idea cannot work.  The only thing that's possible
is to iterate through the whole list of macros.

So since I was iterating through the list, then I decided I didn't want
duplicates.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>