units: move checks to check_ file and hide under the --spammy option
The checks are supposed to be in a check_*.c file instead of a smatch_*.c
file.  These checks are low quality and I almost deleted them but then
instead I just decided to hide them under the --spammy option for people
to play around with if they want.

Looking at the results one thing which I see is if you have an array of
bytes then probably the size should always be in byte units instead of
array_size units...  Something to do.  But for now it's silenced so there
is less urgency.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>