kernel_err_ptr: track error pointers in the database
Originally, error pointers were tracked in a
smatch_data/kernel.returns_err_ptr file.  Also you could kind of infer
that it was an error pointer based on the return values some times.  But
the system was kind of bad.

Better and simpler to just track it in the database.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>