dereference: hard code nla_data() as dereferencing the parameter
The nla_data() does not actually dereference the parameter but it does
dangerous pointer math and something is going to dereference the result
so count it as a dereference.

This finds the bug it was designed to find (and a couple more instances in
the same file) without introducing any false positives in my .config.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>