dereference: fix some false positives
Smatch was counting "fp = &bp->fp[i];" as a dereference of "bp->fp" but
it's not.  It's just pointer math.

This still feels like I'm doing something wrong with how complicated this
code is...  But at least now it's in one place instead of every test
having to implement this.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>