points_to_user_data: update with some ntfs user data functions
The ntfs_read_run_nb() and ntfs_read_bh() functions return user data.  The
ntfs_read_bh() function is kind of ugly because we pass a pointer to the
first member of a union but we end up modifying the whole union.  But it's
still a one liner to handle that so no worries.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>