strlen: store strlen information as a range
I had previously mentioned that strlen() should record the possible
values of strlen() instead of just the highest value.  Mostly strlen
values are ranges, but the common case where they are not is when you
take the strlen of a string constant:

	if (strncmp(p, "foo", strlen("foo")) == 0)

This change makes the strlen data be an rl.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>