flow: handle declarations inside for loops
Somehow I expected that if you had code like:

	for (int i = 0; i < 10; i++) {

That the "int i = 0;" would show up as a declaration statement in the
stmt->iterator_pre_statement but it turns out it's actually stored in the
stmt->iterator_syms.

This didn't matter for the kernel until fairly recently because we never
used to allow these kinds of declarations.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>