db/fixup_kernel.sh: delete kfree() poison information
Depending on your .config kfree() can poison the data.  However in terms
of flow analysis this is not very useful information to track because say
you want to know "where is "(struct foo)->member" set and what is it set
to?  Then every kfree() means that it's set to unknown.

So everything struct member value is unknown if the struct member is part
of a struct which gets freed.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>