function_hooks: prevent a crashing bug parsing fake calls
This was crashing in a kind of unpredictable way and I can't fully explain
what I was seeing.  I think this is a use after free, but why is the data
still on the list and I thought that wasn't supposed to be freed until we
had finished parsing the outer function...

The fake_calls list is supposed to be function data.  So that's an easy
fix and it fixes the crash.  Also just from review it seems like
parse_fake_calls() wasn't getting called if we don't have the cross
function database.  This doesn't affect the crash, but I think it is
the correct behavior.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>