kernel_netdev_priv: track netdev priv data
How net devices work is you allocate a netdevice but then you specify the
amount of private data that you need for the driver and that gets allocated
on the end of the netdevice struct.  It's all allocated as one big buffer
but you need to call "my_data = netdev_priv();" to access your data.

Smatch wants to track what size a buffer is and what the struct members
are etc.  But this format is not normal so Smatch doesn't know how to
handle it.

What we can do is re-use the container_of param mapper code to and create
a new format which tracks netdev_priv().

I wrote this code to warn about a specific CVE that someone reported.  I
did manage to generate a warning for the CVE, but didn't find anything
else.  I don't think there was any major issues with this code either.
At the time, I was quite worried this would make the database too large,
but it's fine.  And recently I fixed a bug with the too common functions
and that made the DB smaller again.

I wrote this code some time back.  Hopefully, I've merged all the required
bits.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>