flow: add support for __cleanup__ functions
This adds support of handling __cleanup__ functions.  It creates function
hooks that are called at the end of the scope.

The one change is that I changed it so scope hooks are called in LIFO
order.  This doesn't affect existing users.  This shouldn't really have
been needed if I had written the parse.c code correctly to create new
scopes for every __cleanup__.  Oh well.  It doesn't hurt either and it
feels correct.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>