conditions: fix handling of "foo = bar ?: baz;"
Handling the "foo = bar ?: baz;" assignment is quite tricky because of the
types and the side effects.  What this code was doing before was:

	foo = bar;
	if (!foo)
		foo = baz;

This has two problems, because checking foo is not the same and checking
bar if the types are different.  Also it evaluates "foo" twice.  Now
what the code does is:

	fake_assign = bar;
	if (fake_assign)
		foo = fake_assign;
	else
		foo = baz;

This fixes the type problem and the double evaluate problem.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>