flow: introduce shallow assignments
I quite like the idea of creating fake assignments but one of the recurring
issues that I have is when assignments have side effects.  So I have the
idea that maybe we can create shallow assignments that have no side
effects.

This is what this patch does.

I'm not sure, I have fully understood all the issues...  But also I have
been running this code for some time and at this time I don't want to have
this diff with the published code.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>