function_hooks: add add_function_param_hook()
I like the idea of a param_key table, but it only works for name/sym
pairs.  It doesn't work for constants.  So what this does specify the
param and it gives you the expression.

Writing this commit message, I think probably a better approach would be
to have the param/key hook but it fakes up an expression pointer for you.
The code to fake up expressions isn't complete or the greatest code but
you could improve that as you need to.

Anyway, that's for a future date.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>