function_hooks: change add_function_param_hook() to add_param_key_expr_hook()
While I was writing the commit message to an earlier commit this morning
I realized that actually we will eventually want to pass a key and fake
an expression.  So change the code to do that.

Otherwise, I will create the new code and then have to update all the users
to the more flexible and powerful code later.  Which I won't do.  But will
instead leave it as-is and hate the mess of it.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>