kernel: use hard coded returns from IS_ERR() for u-boot
This is for parsing u-boot code with no cross function database.
The IS_ERR() function is implemented slightly different on u-boot and
parsing it inline doesn't work the same way as in the kernel.

This code assumes that when we have implications then IS_ERR() is just
parsed correctly without doing anything special.  Which is not true on
u-boot because of how the inlining works.

It looks like this code was written to handling an ordering issue with
regards to implications.  But so far as I can see the ordering issue has
been fixed so we can just delete it.  Implications seem work fine and I
haven't seen any negative results from testing this code on the kernel.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>