refcount_info/free: fix kfree_skb() when there is no DB
The kfree_skb() function does not actually free data until we drop the
last reference.  This is checked by check_kfree_strict.c already and it
works when you have the cross function database but it does not work if
you don't.

There are two issues:
1) The skb_get() function needs to be hard coded as a refcount inc
   function.  If you have the DB this will be handled automatically, but
   it's such an important function that we need to hard code it as well.
2) There is an ordering bug which meant that kfree_skb() was recorded as
   a refcount dec function before we checked whether the reference count
   was incremented in check_kfree_strict.c.  Use the _late() hook to fix
   this ordering issue.

When I ran with this code, it generated about 10 false positives to do
with kref_put().

	kref_put(&dev->count, dev_free);
	kref_put(&dev->count, dev_free);

Obviously calling dev_free() twice is a bug, but it's only called for the
second _put().  What I'd like to do is change the first call to:

	kref_put(&dev->count, warn_if_called);

That will silence the static checker warning and add some debugging output
to potentially help identify bugs at runtime.  Maybe someone will write
this code for me?

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>