buf_size: big ugly sync with released sources
I've been running different code from the released code for a long time
and now I've forgotten which bits are important or why the code is the
way it is.

The most ugly part of this is in get_array_size_bytes_rl() where I call
filter_size_rl(ret) and have some dead code...  Sorry!

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>