preempt_info: add class_raw_spinlock_irqsave_destructor()
Smatch thinks that class_raw_spinlock_irqsave_constructor() always
disables preempt but that class_raw_spinlock_irqsave_destructor() only
conditionally restores it.  The if statement looks like this:

	if (_T->lock) { _unlock; }

Hard code that it always restores the preempt count.  This is sort of a
not ideal.  Ideally this function would just be parsed inline...

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>