dereference: handle frob(&foo->bar); better
When you call frob(&foo->bar); then Smatch doesn't necessarily mark it as a
dereference because it isn't.  It's just passing an address.  But say
frob(); dereferences $0 then that should be handled as a dereference.  So
we have to do that here.

I initially thought maybe this should be handled in smatch_flow.c but
smatch_flow.c doesn't have the DEREFERENCES_PARAM information so it has to
be handled here.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>