atomic_dec_test_path: hack around assignments
Normally this code looks like:

	if (refcount_dec_and_test(ref))
		free();

But sometimes what happens is test people do:

	done = refcount_dec_and_test(ref);
	if (done)
		free();

The right way to fix this is to probably make a list of dec_and_test()
functions and that true means done.  But instead of that, we're just going
to consider everything after the refcount_dec_and_test() on the
dec_and_test() path regardless of the return value.  This will probably
work fine in real life.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>