double_checking: disable implications
The double checking warning is supposed to detect puzzling/pointless
checks:

	ret = frob();
	if (ret)
		return ret;

	frob();
	if (ret)
		return ret;

See the second frob() should be "ret = frob();".  But what sometimes
happens is that if you have implications turned on it warns "Well you
see these variables and these variables are connected so if you pass x to
function foo() then it means 'ret' is zero.  No need to check!"  Which is
sometimes true but it's not useful information.

Just turn off implications.  It's so weird to try make Smatch dumber but
that's where we are.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>