deref: fix a crashing bug
It took me a while to notice this because I use a different check instead
of check_deref.c...  I should just delete check_deref.c.

Anyway, the issue with scoped states is that it frees the memory at the end
of the scope.  Which was fine for normal scopes, but when we started
considering return statements as the end of a scope then you could end up
with several free statements in a scope (double free).

This idea of scoped states was something that seemed important but I never
ended up going with it.  It could still be useful for saving memory and
cleaner debug output so I might want to revisit it, but not like this and
without the double free.  For now, just delete this useless code.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>