kernel_user_data: don't insert useless data into the DB
The estate_assigned() stuff was supposed to limit the data being printed
but it ended up printing stuff that wasn't allowed before.  So in this
case we have:

	struct something *p = argument;

The p->foo is "assigned" but the param was not actually set.  We both
tests.

So far as I can see this doesn't cause any problems, but it does add
confusing data to the DB and it does waste resources.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>