check_ida_alloc: Improve check to not warn when min is '1'
This check warns when the max argument of ida_alloc_range is not of the
the form "2^x-1".

It is fine to not warn when min argument is 1 and max argument is a
power of 2.

Before this patch:
drivers/net/ethernet/mellanox/mlx5/core/lib/macsec_fs.c:587 macsec_fs_tx_setup_fte() error: Calling ida_alloc_range() with a 'max' argument which is a power of 2. -1 missing?

After this patch:
Smatch will not warn as the min argument is one.

Signed-off-by: Harshit Mogalapalli <harshit.m.mogalapalli@oracle.com>
Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>