unchecked_allocation: use the smatch_dereference.c hook
This wasn't transitioned to using smatch_dereference.c.  The transition is
a clean up, but it also fixes an ordering bug handling dereferenced
parameters.  Before the smatch_extra.c code was run first and marked the
parameter as non-NULL before this code triggered a warning.

The smatch_dereference.c code may fix a few other more subtle bugs as well.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>