locking: handle class_mutex_destructor()
I don't know how this passed my testing initially.  It should have been
"*$" instead of "$".

But here I've just created a custom function to parse it.  The trick is
that we pass &scope to class_mutex_destructor() but the lock we want is
what was assigned to "scope" (no ampersand).

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>