db/kernel.return_fixes: fix scope_guard() parsing
The scope_guard() macro looks something like this:

scoped_guard (rcu) {
	...
}

The problem is that we recently added support for try_locks where if we
fail to take the lock then we don't enter the block.  However for ordinary
lock functions we do enter the block.  The way we differentiate is by
testing the lock_ptr.

Unfortunately, Smatch doesn't know that the lock_ptr is always non-NULL for
ordinary locks.  This leads to uninitialized variable warnings.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>