flow: fake returns for inline functions
When you have a void function there isn't necessarily a returns statement
at the end.  In fact in the kernel, there never is because checkpatch will
complain.  However this is a like a weird edge case, it would be easier
if there was always a return.

I thought that was how Smatch was working, but apparently I forgot to
implement that for inline functions.

The way that this showed up was because the ordering was wrong with regards
to handling scope_guards.  It should be that we parse the scope_guards
first so that we unlock or whatever, and then insert the return_data into
the database.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>