flow/conditions: fix ordering bug
With the scoped_guard code we get stuff that looks like:

	scoped_guard(cred, old_cred)
		err = ovl_dir_read_merged(dentry, list, &root);
	if (err)
		return err;

The scoped_guard() macros work for locking function which always take the
lock and locking functions which like try_lock() which may or may not
succeed.  When you pre-compile the scope guard macro it looks something
like this (different example):

for (class_mutex_t scope __attribute__((__cleanup__(class_mutex_destructor))) =
	class_mutex_constructor(&chip->lock), *done = 0);
     class_mutex_lock_ptr(&scope) && !done;
     done = 1) {
	...
}

So for most locks it goes through the loop exactly one time, but for try
locks it goes through the loop either zero times or one time.  That is
determined by the value of class_mutex_lock_ptr(&scope).  It's really
important to know the return of class_mutex_lock_ptr(), otherwise in the
first example, Smatch will warn that "err" might be uninitialized because
we didn't necessarily enter the loop.

The problem here is an issue of ordering.  The class_mutex_lock_ptr() is
parsed as an inline.  But the pre_loop code was looking up the return
values before it had been parsed.  We also cannot look up the return value
after because that's the value for the true path.  We have to look it up
at the time when it's parse and that has to be done inside
smatch_conditions.c.  The final complication is that we have to set the
confidence level.  When we're parsing if statements we treat always true
statements different that's a "known true" statement, but with loops we
consider it as entering the loop when it's an "implied true" statement.

Thinking about it now, I re-wrote split_if_statement() to fix the ordering
bug, but it probably wasn't necessary.  For known true/false conditions
probably the ordering doesn't matter.  (As opposed to implied true/false
where it does).  However, it feels nice to write this in the correct way,
and I want to push this further because I think there might be an issue
handling known t/f conditions in compound statements.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>