kernel_irq_context: track when we're in IRQ context
This code is sort of like tracking preempt disabled except even easier.
With preemption, functions can enable and disable it but with IRQ context
it's just you're either an IRQ handler or you're not.

This code isn't properly tested, and the published code doesn't print any
warnings.  But when I ran it before it printed just over 200k warnings...
LOL.

Why am I publishing it now?  Since then I've made two improvements to the
code.  First if we sleep that means we're not in IRQ context so call
clear_irq_context() for that.  Second, I've added a few more function
pointers to the is_ignored_fn().  The is_ignored_fn() list is kind of a
big hammer approach.  I thought about making it granular so it would
ignore one call site at a time, but then it was like actually, "Nah, big
hammers are the coolest type of hammer."

I really think the approach of just adding more function pointers to
the is_ignored_fn() list is going to work so I'm publishing this code.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>