irq_context: ignore (struct flexcop_device)->read_ibi_reg
I think that the situation is that flexcop USB and PCI drivers share a lot
of code but the PCI IRQ handler will not call the USB read/write handler.
So even though these function pointers have the same name the callers
know which they are using and treat them differently.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>