function_hooks: add missing return states
There seems to be some kind of mismatch here where fake_a_param_assignment()
seems to think we could fake a parameter assignment, but then we pop
the ->fake_param_assign_stack and there wasn't any assignment set.  The
truth is that I didn't really investigate why that mismatch exists.

Instead, I just said that if we don't have an assignment, then use the
returned value like we normally would.  This is a good thing to do anyway.
Although it might have been interesting to do that investigation and
possibly there are additional changes we could make as well?  Who knows.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>