mtag: fix a bug on 32 bit systems
When Smatch is creating mtags it was working on u64 values, but it
should be working on unsigned long values.  Otherwise what happens is
that the caller is like, "this is an invalid pointer, let's just say we
don't know what this function can return".  This leads to confusing error
pointer false positives on 32 bit systems.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>