kernel_xa_err: handle xa_is_err() and xa_err() a little bit better
Ideally, xa_err() would be handled like error pointers, but that is kind of
a lot of work.  For example, I massage the output of error pointers so
they are displayed as negatives.  Ugh... As I'm writing this, I realize
that probably I will need to change ptr_max to be lower to accommodate
xa_errors...  I already wish I had more bits to handle mtags on 32bit
systems...  *sigh*  The other thing which is different about pointer errors
is that they're just casts so they're easier to handle instead of the
"(err << 2) | 2" transforms that xa_err() does.

So this patch is really just a starting point for handling xa_err().  It's
not as widely used as error pointers so I don't know how much work I'm
going to have to do to make this work adequately...

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>