helper: fix crashing bug in token_to_ul()
The token_to_ul() function is really rubbish...  It basically lets you
parse a define which is a constant but has some parentheses around it.
I'm not sure exactly how parsing constants is supposed to be done.
I'm going to have to figure that out because the new GFP_ defines are
BIT() calls instead of simple constants.

But for now just stop the token_to_ul() function from crashing due to
recursion.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>